+++ src/confluent_kafka/_model/__init__.py
@@ -149,17 +149,3 @@ class IsolationLevel(Enum):
         if self.__class__ != other.__class__:
             return NotImplemented
         return self.value < other.value
-
-
-class DeletedRecords:
-    """
-    DeletedRecords
-    Represents information about deleted records.
-
-    Parameters
-    ----------
-    low_watermark: int
-        The "low watermark" for the topic partition on which the deletion was executed.
-    """
-    def __init__(self, low_watermark):
-        self.low_watermark = low_watermark
